Added API's for college registration and retrieval

+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/state_medical_council/StateMedicalCouncil.java
@@ -22,7 +22,7 @@ public class StateMedicalCouncil extends CommonAuditEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private BigInteger id;
     private String code;
     private String name;
     private String nameShort;
