+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/user_detail/UserDetailRepository.java
@@ -1,7 +1,9 @@
 package in.gov.abdm.nmr.db.sql.domain.user_detail;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailRepository extends JpaRepository<UserDetail, Long> {
+public interface UserDetailRepository extends JpaRepository<UserDetail, BigInteger> {
 
 }
